+++ include/jemalloc/internal/jemalloc_internal_includes.h
@@ -41,16 +41,12 @@
 /******************************************************************************/
 
 #include "jemalloc/internal/arena_types.h"
-#include "jemalloc/internal/tcache_types.h"
-#include "jemalloc/internal/prof_types.h"
 
 /******************************************************************************/
 /* STRUCTS */
 /******************************************************************************/
 
-#include "jemalloc/internal/prof_structs.h"
 #include "jemalloc/internal/arena_structs.h"
-#include "jemalloc/internal/tcache_structs.h"
 
 /******************************************************************************/
 /* EXTERNS */
@@ -59,9 +55,9 @@
 #include "jemalloc/internal/jemalloc_internal_externs.h"
 #include "jemalloc/internal/arena_externs.h"
 #include "jemalloc/internal/large.h"
-#include "jemalloc/internal/tcache_externs.h"
+#include "jemalloc/internal/tcache.h"
 #include "jemalloc/internal/malloc_dispatch.h"
-#include "jemalloc/internal/prof_externs.h"
+#include "jemalloc/internal/prof.h"
 #include "jemalloc/internal/background_thread.h"
 
 /******************************************************************************/
